the nightmare before presentation

+++ TestProject2/Assets/Scripts/Destructible.cs
@@ -1,21 +1,23 @@
 using UnityEngine;
 using System.Collections;
+using Unity.AI.Navigation;
 
 public class Destructable : MonoBehaviour, IDamage
 {
 
-    [SerializeField] int HP;
+    [SerializeField] float HP;
     [SerializeField] Renderer model;
 
     Color colorOrig;
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
 
         HP -= amount;
 
         if (HP <= 0)
         {
+            //NavMeshSurface
             Destroy(gameObject);
 
         } else
@@ -39,7 +41,6 @@ public class Destructable : MonoBehaviour, IDamage
 
     IEnumerator FlashWhite()
     {
-
         model.material.color = Color.white;
         yield return new WaitForSeconds(0.1f);
         model.material.color = colorOrig;
